+++ src/components/themes/BaseTheme.h
@@ -223,6 +223,7 @@ class BaseTheme {
                              const int pageCount, std::string title, const int paddingBottom = 0,
                              const int textYOffset = 0, const bool isPageBookmarked = false,
                              const char* timeLeftLabel = nullptr) const;
+  virtual void drawTopStatusBarClock(GfxRenderer& renderer) const;
   virtual void drawHelpText(const GfxRenderer& renderer, Rect rect, const char* label) const;
   virtual void drawTextField(const GfxRenderer& renderer, Rect rect, const int textWidth, bool cursorMode = false,
                              int contentStartX = 0, int contentWidth = 0) const;